Moved transformRect to rect extension and added docs
