+++ packages/flame/lib/src/extensions/rect.dart
@@ -3,6 +3,7 @@ import 'dart:math' as math;
 import 'dart:ui';
 
 import '../../geometry.dart';
+import 'matrix4.dart';
 import 'offset.dart';
 import 'size.dart';
 import 'vector2.dart';
@@ -50,6 +51,20 @@ extension RectExtension on Rect {
       bottomRight.toVector2(),
       bottomLeft.toVector2(),
     ];
+  }
+
+
+  /// Transform Rect using the transformation defined by [matrix].
+  /// 
+  /// **Note:** Rotation matrices will increase the size of the [Rect] but they 
+  /// will not rotate it as [Rect] does not have any rotational values.
+  /// 
+  /// **Note:** Only non-negative scale transforms are allowed, if a negative 
+  /// scale is applied it will return a zero-based [Rect].
+  Rect transform(Matrix4 matrix) {
+    final topLeft = matrix.transform2(this.topLeft.toVector2());
+    final bottomRight = matrix.transform2(this.bottomRight.toVector2());
+    return Rect.fromLTRB(topLeft.x, topLeft.y, bottomRight.x, bottomRight.y);
   }
 
   /// Creates bounds in from of a [Rect] from a list of [Vector2]
